+++ sklearn/cluster/tests/test_spectral.py
@@ -24,14 +24,16 @@ def test_spectral_clustering():
                   [0, 0, 0, 2, 2, 3, 2],
                   [0, 0, 0, 1, 3, 1, 4],
                   [0, 0, 0, 1, 2, 4, 1],
-                 ])
+                  ])
 
-    for eig_solver in ('arpack', 'lobpcg'):
+    for eigen_solver in ('arpack', 'lobpcg'):
         for assign_labels in ('kmeans', 'discretize'):
             for mat in (S, sparse.csr_matrix(S)):
                 model = SpectralClustering(random_state=0, n_clusters=2,
-                        affinity='precomputed', eig_solver=eig_solver,
-                        assign_labels=assign_labels).fit(mat)
+                                           affinity='precomputed',
+                                           eigen_solver=eigen_solver,
+                                           assign_labels=assign_labels
+                                           ).fit(mat)
                 labels = model.labels_
                 if labels[0] == 0:
                     labels = 1 - labels
@@ -40,9 +42,9 @@ def test_spectral_clustering():
 
                 model_copy = loads(dumps(model))
                 assert_equal(model_copy.n_clusters, model.n_clusters)
-                assert_equal(model_copy.eig_solver, model.eig_solver)
+                assert_equal(model_copy.eigen_solver, model.eigen_solver)
                 assert_array_equal(model_copy.random_state.get_state()[1],
-                            model.random_state.get_state()[1])
+                                   model.random_state.get_state()[1])
                 assert_array_equal(model_copy.labels_, model.labels_)
 
 
@@ -59,7 +61,7 @@ def test_spectral_lobpcg_mode():
     D = pairwise_distances(X)  # Distance matrix
     S = np.max(D) - D  # Similarity matrix
     labels = spectral_clustering(S, n_clusters=len(centers),
-                                     random_state=0, eig_solver="lobpcg")
+                                 random_state=0, eigen_solver="lobpcg")
     # We don't care too much that it's good, just that it *worked*.
     # There does have to be some lower limit on the performance though.
     assert_greater(np.mean(labels == true_labels), .3)
@@ -84,13 +86,14 @@ def test_spectral_amg_mode():
         amg_loaded = False
     if amg_loaded:
         labels = spectral_clustering(S, n_clusters=len(centers),
-                                     random_state=0, eig_solver="amg")
+                                     random_state=0, eigen_solver="amg")
         # We don't care too much that it's good, just that it *worked*.
         # There does have to be some lower limit on the performance though.
         assert_greater(np.mean(labels == true_labels), .3)
     else:
         assert_raises(ValueError, spectral_embedding, S,
-                      n_components=len(centers), random_state=0, eig_solver="amg")
+                      n_components=len(centers),
+                      random_state=0, eigen_solver="amg")
 
 
 def test_spectral_unknown_mode():
@@ -106,7 +109,7 @@ def test_spectral_unknown_mode():
     S = np.max(D) - D  # Similarity matrix
     S = sparse.coo_matrix(S)
     assert_raises(ValueError, spectral_clustering, S, n_clusters=2,
-                  random_state=0, eig_solver="<unknown>")
+                  random_state=0, eigen_solver="<unknown>")
 
 
 def test_spectral_clustering_sparse():
@@ -122,12 +125,12 @@ def test_spectral_clustering_sparse():
                   [0, 0, 0, 0, 1, 3, 3, 1, 2, 4],
                   [0, 0, 0, 0, 1, 3, 3, 2, 1, 4],
                   [0, 0, 0, 0, 1, 2, 4, 4, 4, 1],
-                 ])
+                  ])
 
     S = sparse.coo_matrix(S)
 
     labels = SpectralClustering(random_state=0, n_clusters=2,
-            affinity='precomputed').fit(S).labels_
+                                affinity='precomputed').fit(S).labels_
     if labels[0] == 0:
         labels = 1 - labels
 
@@ -136,10 +139,10 @@ def test_spectral_clustering_sparse():
 
 def test_affinities():
     X, y = make_blobs(n_samples=40, random_state=1, centers=[[1, 1], [-1, -1]],
-            cluster_std=0.4)
+                      cluster_std=0.4)
     # nearest neighbors affinity
     sp = SpectralClustering(n_clusters=2, affinity='nearest_neighbors',
-            random_state=0)
+                            random_state=0)
     labels = sp.fit(X).labels_
     assert_equal(adjusted_rand_score(y, labels), 1)