+++ doc/whats_new.rst
@@ -72,9 +72,15 @@ Changelog
    - New kernel :class:`metrics.chi2_kernel` by `Andreas Müller`_, often used
      in computer vision applications.
 
+<<<<<<< HEAD
    - Longstanding bug in :class:`naive_bayes.BernoulliNB` fixed by
      Shaun Jackman.
 
+=======
+   - New estimator :class:`manifold.SpectralEmbedding` and function
+     :func:`manifold.spectral_embedding`, implementing the
+     "laplacian eigenmaps" for nonlinear dimensionality reduction by Wei Li.
+>>>>>>> RENAME rename eigen_tol and eigen_solver, and warning about using old variable name eig_tol and mode
 
 API changes summary
 -------------------
@@ -135,6 +141,12 @@ API changes summary
    - The attribute ``gmm`` of :class:`hmm.GMMHMM` was renamed to ``gmm_``
      to adhere more strictly with the API.
 
+  - :func:`cluster.spectral_embedding` is now in
+    :func:`manifold.spectral_embedding`
+
+  - Renamed ``eig_tol`` in :func:`manifold.spectral_embedding`,
+    :class:`cluster.SpectralClustering` to ``eigen_tol``, renamed ``mode``
+    to ``eigen_solver``
 
 .. _changes_0_12.1: